Update dropbox.js
Eliminated filetype restriction from the Dropbox FilePicker on V12. This assumes that the user is knowledgeable enough to use the appropriate file type for the appropriate application.

Only the FilePicker function is used in V12. The other functions (TileConfig, audio, folder selection, etc) do not seem to be used, at least on V12, so those functions are untouched.